refactor: extract shared utilities
